merge

+++ app/src/main/java/com/egoriku/catsrunning/ui/statisticChart/FitChart.java
@@ -82,13 +82,13 @@ public class FitChart extends View {
     }
 
 
-    public void setMaxValue(float value) {
+    public void setMaxValue(int value) {
         maxValue = value;
     }
 
 
     //single value
-    public void setValue(float value) {
+    public void setValue(int value) {
         chartValues.clear();
         FitChartValue chartValue = new FitChartValue(value, valueStrokeColor);
         chartValue.setPaint(buildPaintForValue());
